requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
duckdb>=1.0.0
pyarrow>=14.0.0
streamlit>=1.32.0
plotly>=5.20.0
